+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Version header format: `[version] Name - year-month-day (entropy-core compatibil
   - new: './src/faucet/command.ts' - main entrypoint for CLI flows
   - new: package added - yocto-spinner for adding loading spinners to the cli
   - new: added new menu item for TUI to trigger a jumpstart to the network (needs to only be run once)
+- new: option to display cli and core version
 
 ### Changed
 
@@ -58,6 +59,7 @@ Version header format: `[version] Name - year-month-day (entropy-core compatibil
 - removed flow/entropyFaucet/*.ts directory with file restructure
 - added faucet to main menu for TUI
 - updated faucet to use loading spinner to indicate to user the progress of the transfer
+- ascii art print out now shows up to date core version based, coming from SDK
 
 ### Broke